fix: include n in list of factors
I noticed this problem when trying to input a prime as a number - the GCF of 3 and 15 should be 3, not 1! Thankfully, that check I thought was redundant caught this whereas the Test object didn't. This commit also fixes an issue from c249375, needed to add an else statement to make sure not to append negative integers, not just warn about doing it in the elif statement.